Chore : BreedConverter 삭제 및 pet 몸무게 실수형 변환

+++ api/src/main/java/com/petid/api/pet/dto/PetAppearanceDto.java
@@ -7,7 +7,7 @@ public record PetAppearanceDto() {
             Long appearanceId,
             String breed,
             String hairColor,
-            Integer weight,
+            Double weight,
             String hairLength
     ) {
         public PetAppearance toDomain() {
@@ -25,7 +25,7 @@ public record PetAppearanceDto() {
             Long appearanceId,
             String breed,
             String hairColor,
-            Integer weight,
+            Double weight,
             String hairLength
     ) {
         public static Response from(PetAppearance petAppearance) {
